[VMware to KVM] Cleanup leftover migrated volumes in case of migration failures (#13151)
